Move FlappyLogger to utils.
