+++ lib/src/services/parsing/file_parser.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:meta/meta.dart';
 
-import '../flappy_logger.dart';
+import '../../utils/flappy_logger.dart';
 
 /// A base file parser which should be extended by supported file times
 abstract class FileParser {
